+++ webapp/src/main/java/com/s_giken/training/webapp/service/MemberService.java
@@ -17,11 +17,11 @@ import com.s_giken.training.webapp.model.entity.Member;
 public interface MemberService {
     public List<Member> findAll();
 
-    public Optional<Member> findById(int memberId);
+    public Optional<Member> findById(Long memberId);
 
     public List<Member> findByConditions(MemberSearchCondition memberSearchCondition);
 
     public void save(Member member);
 
-    public void deleteById(int memberId);
+    public void deleteById(Long memberId);
 }
